:recycle: (web) encapsulate reference display values logic
