+++ apps/web/components/field-inputs/reference-record-picker.tsx
@@ -1,14 +1,14 @@
 import type { IQueryRecords, ReferenceField } from '@egodb/core'
-import { getSelectedRecordId, useGetForeignRecordsQuery, useGetRecordQuery } from '@egodb/store'
-import type { MultiSelectProps, SelectItem } from '@egodb/ui'
+import { getSelectedRecordId, useGetForeignRecordsQuery } from '@egodb/store'
+import type { MultiSelectProps } from '@egodb/ui'
 import { useDisclosure } from '@egodb/ui'
 import { Loader } from '@egodb/ui'
 import { Group } from '@egodb/ui'
 import { MultiSelect } from '@egodb/ui'
-import { isEmpty } from '@fxts/core'
 import { forwardRef } from 'react'
 import { useAppSelector } from '../../hooks'
 import { useCurrentTable } from '../../hooks/use-current-table'
+import { useReferenceDisplayValues } from '../../hooks/use-reference-display-values'
 import { RecordValue } from '../field-value/record-value'
 import { FieldIcon } from './field-Icon'
 
@@ -31,18 +31,8 @@ export const ReferenceRecordPicker: React.FC<IProps> = ({ field, ...rest }) => {
   const recordId = useAppSelector(getSelectedRecordId)
   const table = useCurrentTable()
   const foreignTableId = field.foreignTableId.into() ?? table.id.value
-  const displayFields = field.displayFieldIds.map((f) => f.value)
 
   const [focused, handelr] = useDisclosure(false)
-  const { data: record } = useGetRecordQuery(
-    {
-      tableId: table.id.value,
-      id: recordId,
-    },
-    {
-      skip: !recordId,
-    },
-  )
 
   const { rawRecords: foreignRecords, isLoading } = useGetForeignRecordsQuery(
     { tableId: table.id.value, foreignTableId, fieldId: field.id.value },
@@ -54,21 +44,7 @@ export const ReferenceRecordPicker: React.FC<IProps> = ({ field, ...rest }) => {
     },
   )
 
-  // TODO: Encapsulation
-  const data: SelectItem[] = []
-  if (record) {
-    const foreignRecordIds = record.values[field.id.value]
-    if (Array.isArray(foreignRecordIds) && !isEmpty(foreignRecordIds)) {
-      const values = field.getDisplayValues(record.displayValues)
-      for (const [index, foreignRecordId] of foreignRecordIds.entries()) {
-        data.push({ value: foreignRecordId as string, label: values[index]?.filter(Boolean).toString() ?? '' })
-      }
-    }
-  }
-  for (const foreignRecord of foreignRecords) {
-    const values = displayFields.map((fieldId) => foreignRecord.values[fieldId]?.toString())
-    data.push({ value: foreignRecord.id, label: values.filter(Boolean).toString() })
-  }
+  const data = useReferenceDisplayValues(field, recordId, foreignRecords)
 
   return (
     <MultiSelect
