+++ apps/web/components/field-inputs/tree-records-picker.tsx
@@ -1,12 +1,12 @@
 import type { IQueryRecords, TreeField } from '@egodb/core'
 import { getSelectedRecordId, useGetRecordQuery, useLazyTreeAvailableQuery } from '@egodb/store'
-import type { MultiSelectProps, SelectItem } from '@egodb/ui'
+import type { MultiSelectProps } from '@egodb/ui'
 import { Group } from '@egodb/ui'
 import { Loader, MultiSelect } from '@egodb/ui'
-import { isEmpty } from '@fxts/core'
 import { forwardRef, useEffect, useState } from 'react'
 import { useAppSelector } from '../../hooks'
 import { useCurrentTable } from '../../hooks/use-current-table'
+import { useReferenceDisplayValues } from '../../hooks/use-reference-display-values'
 import { RecordValue } from '../field-value/record-value'
 import { FieldIcon } from './field-Icon'
 
@@ -28,7 +28,6 @@ const TreeSelectItem = forwardRef<HTMLDivElement, ItemProps>(({ label, ...others
 export const TreeRecordsPicker: React.FC<IProps> = ({ field, ...rest }) => {
   const recordId = useAppSelector(getSelectedRecordId)
   const table = useCurrentTable()
-  const displayFields = field.displayFieldIds.map((f) => f.value)
 
   const [focused, setFocused] = useState(false)
   const [getRecords, { rawRecords: foreignRecords, isLoading }] = useLazyTreeAvailableQuery({
@@ -37,34 +36,12 @@ export const TreeRecordsPicker: React.FC<IProps> = ({ field, ...rest }) => {
       rawRecords: (Object.values(result.data?.entities ?? {}) ?? []).filter(Boolean) as IQueryRecords,
     }),
   })
-  const { data: record } = useGetRecordQuery(
-    {
-      tableId: table.id.value,
-      id: recordId,
-    },
-    {
-      skip: !recordId,
-    },
-  )
 
   useEffect(() => {
     getRecords({ tableId: table.id.value, treeFieldId: field.id.value, recordId })
   }, [focused])
 
-  const data: SelectItem[] = []
-  if (record) {
-    const foreignRecordIds = record.values[field.id.value]
-    if (Array.isArray(foreignRecordIds) && !isEmpty(foreignRecordIds)) {
-      const values = field.getDisplayValues(record.displayValues)
-      for (const [index, foreignRecordId] of foreignRecordIds.entries()) {
-        data.push({ value: foreignRecordId as string, label: values[index]?.filter(Boolean).toString() ?? '' })
-      }
-    }
-  }
-  for (const foreignRecord of foreignRecords) {
-    const values = displayFields.map((fieldId) => foreignRecord.values[fieldId]?.toString())
-    data.push({ value: foreignRecord.id, label: values?.filter(Boolean).toString() })
-  }
+  const data = useReferenceDisplayValues(field, recordId, foreignRecords)
 
   return (
     <MultiSelect
